Display progress bar based on reps, not discarded cards
Now we fill the bar based on the percentage of total reps present in
the deck, taking jokers into account (jokers act as the previous
card's reps). As a bonus, we display a green region at the tip of the
progress bar indicating how many reps you just completed, which we
only know because we can look at the discard pile.