Keep deck intact in the state and use indexes to find draw, discard
We still reconstruct the "draw" and "discard" arrays by slicing out of
the main deck array because so many parts of the program rely on those
existing, but we may do away with them eventually.

Mainly, this does not mutate the deck itself as you go through the
workout, instead it moves a draw index backwards across the array.