Add help page, move toward style standard
We are going to use single quotes in JS and double quotes in JSX
starting from now. Also we're going to do away with the containers and
instead merge the component classes and Redux wiring into a single
file and export the wired up version without this "Visible" prefix
garbage.