Move deck share string into the hash
This resolves a (major) bug where the app won't load at all when
visiting a shared link unless the app was previously cached by your
browser. Because the server can't route a "/deadbeef" URL, it would
404 instead of loading the app and letting Javascript route it.

If the page was already cached, the cached Javascript would take over
and handle the route.