Reverse additions to the discard pile
Because we draw from the end of the deck array, and because slice()
returns the cards in their reversed order, we actually want to display
them re-reversed. This only applies to drawing three cards at a time.

The net effect is that the actual "top" card is the first card
displayed.